requirements.txt
bson
pandas
wtforms
easysnmp
dash-core-components
dash-html-components
dash-table
dash
flask-wtf
flask
PyYAML
pymongo
